ISPN-11156 Move the CustomFailurePolicy interface to a different package than org.infinispan.xsite
