+++ core/src/main/java/org/infinispan/xsite/CustomFailurePolicy.java
@@ -1,14 +1,5 @@
 package org.infinispan.xsite;
 
-import java.util.Collection;
-import java.util.Map;
-import java.util.function.BiFunction;
-import java.util.function.Function;
-
-import javax.transaction.Transaction;
-
-import org.infinispan.Cache;
-
 /**
  * Used for implementing custom policies in case of communication failures with a remote site. The handle methods are
  * allowed to throw instances of {@link BackupFailureException} to signal that they want the intra-site operation to
@@ -23,47 +14,9 @@ import org.infinispan.Cache;
  * @author Mircea Markus
  * @see BackupFailureException
  * @since 5.2
+ * @deprecated since 11. Use {@link org.infinispan.configuration.cache.CustomFailurePolicy} instead.
  */
-public interface CustomFailurePolicy<K, V> {
+@Deprecated
+public interface CustomFailurePolicy<K, V> extends org.infinispan.configuration.cache.CustomFailurePolicy<K, V> {
 
-   /**
-    * Invoked during the initialization phase.
-    */
-   void init(Cache<K, V> cache);
-
-   void handlePutFailure(String site, K key, V value, boolean putIfAbsent);
-
-   void handleRemoveFailure(String site, K key, V oldValue);
-
-   void handleReplaceFailure(String site, K key, V oldValue, V newValue);
-
-   default void handleComputeFailure(String site, K key, BiFunction<? super K, ? super V, ? extends V> remappingFunction, boolean computeIfPresent) {}
-
-   default void handleComputeIfAbsentFailure(String site, K key, Function<? super K, ? extends V> mappingFunction) {}
-
-   default void handleReadWriteKeyFailure(String site, K key) {}
-
-   default void handleReadWriteKeyValueFailure(String site, K key) {}
-
-   default void handleWriteOnlyKeyFailure(String site, K key) {}
-
-   default void handleWriteOnlyKeyValueFailure(String site, K key) {}
-
-   default void handleReadWriteManyFailure(String site, Collection<? extends K> keys) {}
-
-   default void handleReadWriteManyEntriesFailure(String site, Map<? extends K, ? extends V> keys) {}
-
-   default void handleWriteOnlyManyFailure(String site, Collection<? extends K> key) {}
-
-   default void handleWriteOnlyManyEntriesFailure(String site, Map<? extends K, ? extends V> key) {}
-
-   void handleClearFailure(String site);
-
-   void handlePutAllFailure(String site, Map<K, V> map);
-
-   void handlePrepareFailure(String site, Transaction transaction);
-
-   void handleRollbackFailure(String site, Transaction transaction);
-
-   void handleCommitFailure(String site, Transaction transaction);
 }
